Improve API requests and modal refresh
